requirements.txt
Django==2.2.9
djangorestframework==3.11.1
django-cors-headers==3.2.1
django-currentuser==0.5.1
Pillow==7.0.0